foodAllergenDectector/requirements.txt
Flask
Flask-WTF
joblib
numpy
pandas
scikit-learn
pytesseract
Pillow
